simplifica estrutura de pastas
