+++ show-do-milhao/src/main/assets/poliedroFundoBranco.png
