+++ show-do-milhao/src/main/assets/retanguloInferior.png
