+++ show-do-milhao/src/main/ui/view/TelaAdicionaPergunta.java
@@ -1,4 +1,4 @@
-package main.view;
+package view;
 
 import javax.swing.*;
 import java.awt.*;
