+++ show-do-milhao/src/main/ui/view/TelaCadastro.java
@@ -1,4 +1,4 @@
-package main.view;
+package view;
 
 import java.awt.*;
 import java.awt.event.*;
