+++ show-do-milhao/src/main/ui/view/TelaHistAdmin.java
@@ -1,12 +1,11 @@
-package main.view;
+package view;
 
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.*;
 
-public class TelaHistorico extends JFrame{
-    public TelaHistorico(Runnable telaMenuEst, Runnable configs){
-        // var rosa = new Color(238, 33, 82);
+public class TelaHistAdmin extends JFrame{
+    public TelaHistAdmin(Runnable telaMenuAdmin, Runnable configs){
         var roxo = new Color(20, 14, 40);
 
         // Tela
@@ -39,7 +38,7 @@ public class TelaHistorico extends JFrame{
         seta.addMouseListener(new MouseAdapter() {
             @Override
             public void mouseClicked(MouseEvent e){
-                telaMenuEst.run();
+                telaMenuAdmin.run();
                 dispose();
             }
         });
